Resolving unicode parsing issue in Delphi.
The test may seem less meaningful this way, but it is still valid.
The code-points used in the
  Test_EncodeCodepointToString and
  Test_DecodeCodepointFromString
test methods are now passed as literal characters. All conversion is still performed, but the tests do not show the characters as they would normally appear in source.
This is unavoidable as some earlier versions of the Rad Studio IDE do not correctly parse unicode literals as characters, causing tests to fail.