Tearing out dynamic log.
Sorry, but this feature was always experimental and required far more maintenance than it was worth. As we're before first official release version, I decided to tear it out to save the head-ache.
If it is necessary to pass log messages out from a module to another, a call-back driven by an event log target is sufficient.